LPS-105694 Move to UserDataFactory
