+++ modules/util/portal-tools-sample-sql-builder/src/main/java/com/liferay/portal/tools/sample/sql/builder/SampleSQLBuilder.java
@@ -264,6 +264,8 @@ public class SampleSQLBuilder {
 							"messageBoardDataFactory",
 							new MessageBoardDataFactory()
 						).put(
+							"userDataFactory", new UserDataFactory()
+						).put(
 							"wikiDataFactory", new WikiDataFactory()
 						).build(),
 						sampleSQLWriter);
