+++ modules/util/portal-tools-sample-sql-builder/src/main/resources/com/liferay/portal/tools/sample/sql/builder/dependencies/default_user.ftl
@@ -1,16 +1,16 @@
 <#-- Default user -->
 
-<@insertUser _userModel=dataFactory.newDefaultUserModel() />
+<@insertUser _userModel=userDataFactory.newDefaultUserModel() />
 
 <#-- Guest user -->
 
-<#assign guestUserModel = dataFactory.newGuestUserModel() />
+<#assign guestUserModel = userDataFactory.newGuestUserModel() />
 
-<@insertGroup _groupModel=dataFactory.newGroupModel(guestUserModel) />
+<@insertGroup _groupModel=userDataFactory.newGroupModel(guestUserModel) />
 
 <#assign
 	groupIds = [guestGroupModel.groupId]
-	roleIds = [dataFactory.administratorRoleId]
+	roleIds = [userDataFactory.administratorRoleId]
 />
 
 <@insertUser
@@ -22,9 +22,9 @@
 <#-- Sample user -->
 
 <#assign
-	sampleUserModel = dataFactory.newSampleUserModel()
+	sampleUserModel = userDataFactory.newSampleUserModel()
 
-	userGroupModel = dataFactory.newGroupModel(sampleUserModel)
+	userGroupModel = userDataFactory.newGroupModel(sampleUserModel)
 
 	layoutModel = dataFactory.newLayoutModel(userGroupModel.groupId, "home", "", "")
 />
@@ -34,8 +34,8 @@
 <@insertGroup _groupModel=userGroupModel />
 
 <#assign
-	groupIds = dataFactory.getSequence(dataFactory.maxGroupCount)
-	roleIds = [dataFactory.administratorRoleId, dataFactory.powerUserRoleId, dataFactory.userRoleId]
+	groupIds = dataFactory.getSequence(userDataFactory.maxGroupCount)
+	roleIds = [userDataFactory.administratorRoleId, userDataFactory.powerUserRoleId, userDataFactory.userRoleId]
 />
 
 <@insertUser
