+++ modules/util/portal-tools-sample-sql-builder/src/main/resources/com/liferay/portal/tools/sample/sql/builder/dependencies/groups.ftl
@@ -1,16 +1,16 @@
-<#assign guestGroupModel = dataFactory.newGuestGroupModel() />
+<#assign guestGroupModel = userDataFactory.newGuestGroupModel() />
 
 <#include "default_user.ftl">
 
 <@insertLayout _layoutModel=dataFactory.newLayoutModel(guestGroupModel.groupId, "welcome", "com_liferay_login_web_portlet_LoginPortlet,", "com_liferay_hello_world_web_portlet_HelloWorldPortlet,") />
 
-<@insertGroup _groupModel=dataFactory.newGlobalGroupModel() />
+<@insertGroup _groupModel=userDataFactory.newGlobalGroupModel() />
 
 <@insertGroup _groupModel=guestGroupModel />
 
-<@insertGroup _groupModel=dataFactory.newUserPersonalSiteGroupModel() />
+<@insertGroup _groupModel=userDataFactory.newUserPersonalSiteGroupModel() />
 
-<#list dataFactory.newGroupModels() as groupModel>
+<#list userDataFactory.newGroupModels() as groupModel>
 	<#assign groupId = groupModel.groupId />
 
 	<#include "asset_publisher.ftl">
